requirements.txt
openvr
glfw
PyOpenGL
Pillow
bilibili-api-python
pywebview
aiohttp